+++ docs/CHANGES.md
@@ -1,6 +1,7 @@
 2017-05-29  Release 1.81.14
 
-  * Make CheckMarx default requirement for applications.
+  * Make CheckMarx a default requirement for Application domain objects.
+  * Generate a default project definition file if no project definition file exists yet.
 
 
 2017-05-26  Release 1.81.13